requirements.txt
# Knowledge Graph Social Network System Dependencies
networkx>=3.1            # For graph data structures and algorithms
fastapi>=0.103.1         # For API development
uvicorn>=0.23.2          # ASGI server for FastAPI
pydantic>=2.3.0          # Data validation and settings management
python-jose>=3.3.0       # JWT token handling for authentication
passlib>=1.7.4           # Password hashing
python-multipart>=0.0.6  # Form data parsing
sqlalchemy>=2.0.20       # ORM for database operations
aiofiles>=23.2.1         # Async file operations
pillow>=10.0.0           # Image processing
numpy>=1.25.2            # Numerical operations
scikit-learn>=1.3.0      # Machine learning algorithms
pandas>=2.1.0            # Data manipulation
matplotlib>=3.7.2        # Data visualization
pytest>=7.4.0            # Testing
pytest-asyncio>=0.21.1   # Async testing support
python-dotenv>=1.0.0     # Environment variable management 
pydantic[email]          # Email validation
jinja2>=3.1.2            # Template engine for frontend
google-genai>=1.4.0      # Google Gemini API for gemini-2.0-flash model
requests>=2.31.0         # HTTP requests
fastapi-jwt-auth>=0.5.0  # JWT authentication for FastAPI
openai>=1.12.0           # OpenAI API for DALL-E-2 integration
